# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Clinical Trial Data Collection Platform API
  description: |
    HIPAA-compliant RESTful API for clinical trial patient data collection platform.
    
    This API provides secure endpoints for:
    - User authentication with JWT and MFA support
    - Patient data management with column-level encryption
    - Questionnaire creation and assignment
    - Secure response collection and storage
    - Media file uploads with virus scanning
    - Comprehensive audit logging
    
    ## Security Features
    - JWT Bearer token authentication
    - Role-based access control (Physician vs Patient)
    - Multi-factor authentication (MFA) support
    - Column-level data encryption for PHI
    - HIPAA-compliant audit logging
    - Rate limiting and request throttling
    - Virus scanning for file uploads
    
    ## Data Privacy
    - All personally identifiable information (PII) is encrypted at rest
    - PHI is never exposed in error messages
    - Comprehensive audit trails for all data access
    - Secure token-based session management
    
    ## Authentication Flow
    1. Login with email/password to receive access token
    2. Include Bearer token in Authorization header for all requests
    3. Refresh tokens before expiration using /auth/refresh
    4. Complete MFA verification if enabled
    
  version: 1.0.0
  contact:
    name: Clinical Trial Platform Development Team
    email: dev@clinicaltrial.platform
  license:
    name: Proprietary
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token obtained from /auth/login endpoint.
        
        Example: Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
  schemas:
...
    PatientResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique response identifier
        questionnaireId:
          type: string
          format: uuid
          description: Associated questionnaire ID
        questionnaireVersion:
          type: string
          description: Version of questionnaire used
          example: "1.2"
        status:
          type: string
          enum: [not_started, in_progress, completed, submitted, reviewed]
          description: Current status of the response
          example: completed
        startedAt:
          type: string
          format: date-time
          description: When the response was started
        completedAt:
          type: string
          format: date-time
          description: When the response was completed
        submittedAt:
          type: string
          format: date-time
          description: When the response was submitted
        reviewedAt:
          type: string
          format: date-time
          description: When the response was reviewed by physician
        autoSavedAt:
          type: string
          format: date-time
          description: Last auto-save timestamp
        responseData:
          type: object
          description: Encrypted response data (only returned to authorized users)
          additionalProperties: true
      required:
        - id
        - questionnaireId
        - questionnaireVersion
        - status
...
paths:
...
  /patients/profile:
    get:
      tags:
        - Patients
      summary: Get patient profile
      description: |
        Retrieve the complete profile information for the authenticated patient.
        
        Includes personal information (decrypted for the patient), medical history,
        and account settings. All PII is encrypted at rest and decrypted for display.
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  profile:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      email:
                        type: string
                        format: email
                      patientId:
                        type: string
                        example: PT-2024-001
                      firstName:
                        type: string
                        example: Jane
                      lastName:
                        type: string
                        example: Doe
                      phone:
                        type: string
                        example: +1-555-0456
                      address:
                        type: string
                        example: 123 Main St, City, ST 12345
                      emergencyContact:
                        type: string
                        example: John Doe - Spouse - +1-555-0457
                      gender:
                        type: string
                        enum: [male, female, other, prefer_not_to_say]
                      status:
                        type: string
                        enum: [active, inactive, completed, withdrawn]
                      enrollmentDate:
                        type: string
                        format: date
                      consentSigned:
                        type: boolean
                      consentDate:
                        type: string
                        format: date-time
                      consentVersion:
                        type: string
                        example: "2.1"
                      mfaEnabled:
                        type: boolean
                      createdAt:
                        type: string
                        format: date-time
    put:
      tags:
        - Patients
      summary: Update patient profile
      description: |
        Update the patient's profile information.
        
        All PII fields are automatically encrypted before storage. Medical history
        and sensitive data require appropriate validation and audit logging.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  minLength: 1
                  maxLength: 50
                lastName:
                  type: string
                  minLength: 1
                  maxLength: 50
                phone:
                  type: string
                  pattern: '^\+?[1-9]\d{1,14}$'
                address:
                  type: string
                  maxLength: 200
                emergencyContact:
                  type: string
                  maxLength: 200
                gender:
                  type: string
                  enum: [male, female, other, prefer_not_to_say]
                medicalHistory:
                  type: string
                  maxLength: 2000
                medications:
                  type: string
                  maxLength: 1000
                allergies:
                  type: string
                  maxLength: 1000
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Profile updated successfully
                  updatedAt:
                    type: string
                    format: date-time
...
  /patients/questionnaires/{questionnaireId}:
    get:
      tags:
        - Patients
      summary: Get questionnaire details with patient's response
      description: |
        Retrieve detailed questionnaire information including questions and any existing
        patient response data.
        
        Response data is decrypted and returned if the patient has started the questionnaire.
        Access permissions are verified before returning data.
      parameters:
        - name: questionnaireId
          in: path
          required: true
          description: Unique questionnaire identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Questionnaire details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  questionnaire:
                    allOf:
                      - $ref: '#/components/schemas/Questionnaire'
                      - type: object
                        properties:
                          response:
                            allOf:
                              - $ref: '#/components/schemas/PatientResponse'
                              - type: object
                                description: Current patient response (null if not started)
        '403':
          description: Access denied to questionnaire
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                error: Access denied
                message: You do not have access to this questionnaire
                code: QUESTIONNAIRE_ACCESS_DENIED
        '404':
          description: Questionnaire not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /responses/{responseId}/submit:
    post:
      tags:
        - Responses
      summary: Submit response for review
      description: |
        Submit a completed response for physician review.
        
        This action:
        - Changes status to 'submitted'
        - Sets submission timestamp
        - Marks completion timestamp if not already set
        - Prevents further modifications
        - Triggers notification to assigned physician
      parameters:
        - name: responseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Response submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Response submitted successfully
                  response:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      status:
                        type: string
                        example: submitted
                      submittedAt:
                        type: string
                        format: date-time
                      completedAt:
                        type: string
                        format: date-time
        '400':
          description: Response already submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                error: Response already submitted
                code: RESPONSE_ALREADY_SUBMITTED
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # Media Upload Endpoints
...